Former-commit-id: 04658cacf79607430cd65b14212973c2e689d070

+++ ide/src/org/lgna/story/resourceutilities/ModelResourceExporter.java
@@ -122,7 +122,9 @@ public class ModelResourceExporter {
 	private List<String> tags = new LinkedList<String>();
 	private AxisAlignedBox boundingBox;
 	private File xmlFile;
+	private File javaFile;
 	private Map<String, Image> thumbnails = new HashMap<String, Image>();
+	private Map<String, File> existingThumbnails = null;
 	
 	private String attributionName;
 	private String attributionYear;
@@ -231,10 +233,27 @@ public class ModelResourceExporter {
 	{
 		this.thumbnails.put(name, thumbnail);
 	}
+
+	public void addExistingThumbnail(String name, File thumbnailFile) {
+		if (thumbnailFile != null && thumbnailFile.exists()) {
+			if (this.existingThumbnails == null) {
+				this.existingThumbnails = new HashMap<String, File>();
+			}
+			this.existingThumbnails.put(name, thumbnailFile);
+		}
+		else {
+			System.err.println("FAILED TO ADDED THUMBAIL: "+thumbnailFile+" does not exist.");
+		}
+	}
 	
 	public void setXMLFile(File xmlFile)
 	{
 		this.xmlFile = xmlFile;
+	}
+
+	public void setJavaFile(File javaFile)
+	{
+		this.javaFile = javaFile;
 	}
 	
 	private org.w3c.dom.Element createBoundingBoxElement(Document doc)
@@ -584,6 +603,9 @@ public class ModelResourceExporter {
 	        if (!name.endsWith("/"))
 	          name += "/";
 	        name = name.substring(root.length());
+			if (name.startsWith("/")) {
+				name = name.substring(1);
+			}
 	        if (name.length() > 0)
 	        {
 		        JarEntry entry = new JarEntry(name);
@@ -612,6 +634,9 @@ public class ModelResourceExporter {
 
 	    String entryName = source.getPath().replace("\\", "/");
 	    entryName = entryName.substring(root.length());
+		if (entryName.startsWith("/") || entryName.startsWith("\\")) {
+			entryName = entryName.substring(1);
+		}
 	    JarEntry entry = new JarEntry(entryName);
 	    entry.setTime(source.lastModified());
 	    target.putNextEntry(entry);
@@ -633,6 +658,18 @@ public class ModelResourceExporter {
 	      in.close();
 	  }
 	}
+
+	private File getJavaCodeDir(String root)
+	{
+		String packageDirectory = getDirectoryStringForPackage(this.classData.packageString);
+		return new File(root+packageDirectory);
+	}
+
+	private File getJavaFile(String root)
+	{
+		String filename = getDirectoryStringForPackage(this.classData.packageString)+this.name+".java";
+		return new File(root+filename);
+	}
 	
 	private File createJavaCode(String root)
 	{
@@ -641,7 +678,7 @@ public class ModelResourceExporter {
 		String javaCode = createJavaCode();
 		System.out.println(javaCode);
 		System.out.println(System.getProperty("java.class.path"));
-		File javaFile = new File(root+packageDirectory+this.name+".java");
+		File javaFile = getJavaFile(root);
 		TextFileUtilities.write(javaFile, javaCode);
 		return javaFile;
 	}
@@ -676,15 +713,27 @@ public class ModelResourceExporter {
         return null;
     }
 	   
-	
+	private File getXMLFile(String root) {
+		if (!root.endsWith("/") && !root.endsWith("\\")) {
+			root += "/";
+		}
+		String resourceDirectory = root + getDirectoryStringForPackage(this.classData.packageString)+ModelResourceExporter.getResourceSubDirWithSeparator();
+        File xmlFile = new File(resourceDirectory, this.name+".xml");
+		return xmlFile;
+	}
+
 	private File createXMLFile(String root)
 	{
-		String resourceDirectory = root + getDirectoryStringForPackage(this.classData.packageString)+ModelResourceExporter.getResourceSubDirWithSeparator();
-        File outputFile = new File(resourceDirectory, this.name+".xml");
+        File outputFile = getXMLFile(root);
         try
         {
-	        if (this.xmlFile != null)
+	        if (this.xmlFile != null && this.xmlFile.exists())
 	        {
+				if (!outputFile.exists())
+	            {
+	            	FileUtilities.createParentDirectoriesIfNecessary(outputFile);
+	                outputFile.createNewFile();
+	            }
 	        	FileUtilities.copyFile(this.xmlFile, outputFile);
 	        	return outputFile;
 	        }
@@ -743,37 +792,106 @@ public class ModelResourceExporter {
         }
         return null;
 	}
-	
+
+	public String getThumbnailPath(String rootPath, String textureName) {
+		if (!rootPath.endsWith("/") && !rootPath.endsWith("\\")) {
+			rootPath += "/";
+		}
+		String resourceDirectory = rootPath + getDirectoryStringForPackage(this.classData.packageString)+ModelResourceExporter.getResourceSubDirWithSeparator();
+		if (textureName == null) {
+			return resourceDirectory+this.name+".png";
+		}
+		else {
+			return resourceDirectory+this.name + "_" + textureName + ".png";
+		}
+	}
+
 	private List<File> createThumbnails(String root)
 	{
 		List<File> thumbnailFiles = new LinkedList<File>();
-		boolean isFirst = true;
-		for (Entry<String, Image> entry : this.thumbnails.entrySet())
-		{
-			String resourceDirectory = root + getDirectoryStringForPackage(this.classData.packageString)+ModelResourceExporter.getResourceSubDirWithSeparator();
-			if (isFirst)
+		List<String> thumbnailsCreated = new LinkedList<String>();
+		boolean gotBase = false;
+		if (this.existingThumbnails != null && !this.existingThumbnails.isEmpty()) {
+			for(Entry < String, File > entry : this.existingThumbnails.entrySet())
 			{
-				File f = saveImageToFile(resourceDirectory+this.name+".png", entry.getValue());
-				if (f != null ){
-					thumbnailFiles.add(f);
+				if (!gotBase) {
+					File thumb = new File(getThumbnailPath(root, null));
+					try {
+						FileUtilities.copyFile(entry.getValue(), thumb);
+						thumbnailFiles.add(thumb);
+					}
+					catch (IOException e) {
+						e.printStackTrace();
+						return null;
+					}
+					gotBase = true;
+				}
+				if (entry.getValue().exists() ){
+					thumbnailFiles.add(entry.getValue());
+					thumbnailsCreated.add(entry.getKey());
+				}
+				else {
+					System.err.println("FAILED TO FIND THUMBNAIL FILE '"+entry.getValue()+"'");
+					return null;
 				}
 			}
-			File f = saveImageToFile(resourceDirectory+this.name+"_"+entry.getKey()+".png", entry.getValue());
-			if (f != null ){
-				thumbnailFiles.add(f);
+		}
+		for(Entry < String, Image > entry : this.thumbnails.entrySet())
+		{
+			if (!thumbnailsCreated.contains(entry.getKey())) {
+				if (!gotBase)
+				{
+					File f = saveImageToFile(getThumbnailPath(root, null), entry.getValue());
+					if (f != null ){
+						thumbnailFiles.add(f);
+					}
+					gotBase = true;
+				}
+				File f = saveImageToFile(getThumbnailPath(root, entry.getKey()), entry.getValue());
+				if (f != null ){
+					thumbnailsCreated.add(entry.getKey());
+					thumbnailFiles.add(f);
+				}
 			}
 		}
         return thumbnailFiles;
 	}
 	
-	public boolean addToJar(String sourceDirectory, String resourceDirectory, JarOutputStream resourceJarStream, JarOutputStream sourceJarStream)
+	public boolean addToJar(String sourceDirectory, String resourceDirectory, JarOutputStream resourceJarStream, JarOutputStream sourceJarStream, boolean rebuildJavaFile, boolean rebuildXmlFile)
 	{
 		if (!sourceDirectory.endsWith("/") && !sourceDirectory.endsWith("\\")) {
 			sourceDirectory += File.separator;
         }
-		if (this.classData != null) {
-			File javaFile = createJavaCode(sourceDirectory);
-			File sourceDir = javaFile.getParentFile();
+		xmlFile = getXMLFile(resourceDirectory);
+		if (rebuildXmlFile || !xmlFile.exists()) {
+			xmlFile = createXMLFile(resourceDirectory);
+		}
+		boolean addResources = false;
+		File resourceDir = null;
+		if (xmlFile != null) {
+			addResources = true;
+			resourceDir = xmlFile.getParentFile();
+			List<File> thumbnailFiles = createThumbnails(resourceDirectory);
+		}
+		else {
+			System.err.println("FAILED TO MAKE XML FILE FOR "+this.getName()+"--NOT ADDING IT TO JARS.");
+			return false;
+		}
+		boolean addClassData = false;
+		File sourceDir = null;
+		boolean shouldAddClassData = this.classData != null;
+		if (shouldAddClassData) {
+			sourceDir = getJavaCodeDir(sourceDirectory);
+			javaFile = getJavaFile(sourceDirectory);
+			if (rebuildJavaFile || !javaFile.exists()) {
+				try {
+					javaFile = createJavaCode(sourceDirectory);
+				}
+				catch (Exception e) {
+					System.err.println("FAILED TO MAKE JAVA FILE FOR "+this.getName()+"--NOT ADDING IT TO JARS.");
+					return false;
+				}
+			}
 			
 			String[] args = new String[]{javaFile.getAbsolutePath(), "-target", "1.5", "-classpath", System.getProperty("java.class.path")};
 			PrintWriter pw = new PrintWriter(System.out);
@@ -783,37 +901,47 @@ public class ModelResourceExporter {
 			{
 				System.out.println("BOOM!");
 			}
+			else {
+				addClassData = true;
+			}
+			
+		}
+		if (addResources && (shouldAddClassData && addClassData)) {
 			try
 			{
 				System.out.println("Adding "+sourceDir);
 				add(sourceDir, sourceJarStream, sourceDirectory, false);
+				System.out.println("Adding "+resourceDir);
+				add(resourceDir, resourceJarStream, resourceDirectory, true);
+				return true;
 			}
 			catch (Exception e)
 			{
+				System.err.println("FAILED ADDING DATA TO JARS: ");
 				e.printStackTrace();
 				return false;
 			}
 		}
-		
-		File xmlFile = createXMLFile(resourceDirectory);
-		File resourceDir = xmlFile.getParentFile();
-		List<File> thumbnailFiles = createThumbnails(resourceDirectory);
-		try
-		{
-			System.out.println("Adding "+resourceDir);
-			add(resourceDir, resourceJarStream, resourceDirectory, true);
-		}
-		catch (Exception e)
-		{
-			e.printStackTrace();
+		else {
+			System.err.println("NOT ADDING "+this.getName()+" TO JARS.");
+			if (!addResources) {
+				System.err.println("FAILED TO MAKE RESOURCES FOR "+this.getName());
+			}
+			if (shouldAddClassData && !addClassData) {
+				System.err.println("FAILED TO MAKE JAVA CODE FOR "+this.getName());
+			}
 			return false;
 		}
-		return true;
 	}
 	
 	public boolean addToJar(String sourceDirectory, String resourceDirectory, JarOutputStream jos)
 	{
-		return addToJar(sourceDirectory, resourceDirectory, jos, jos);
+		return addToJar(sourceDirectory, resourceDirectory, jos, jos, true, true);
+	}
+
+	public boolean addToJar(String sourceDirectory, String resourceDirectory, JarOutputStream jos, boolean rebuildFiles)
+	{
+		return addToJar(sourceDirectory, resourceDirectory, jos, jos, rebuildFiles, rebuildFiles);
 	}
 	
 	public File export(String sourceDirectory, String resourceDirectory, String outputDir)
